backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.12.4
pydantic-settings==2.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
# psycopg2-binary==2.9.9  # Removed for SQLite
# asyncpg==0.29.0  # Removed for SQLite

# YARA
yara-python==4.5.0
PyYAML==6.0.2

# Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Task Queue
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-magic==0.4.27
pefile==2023.2.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.19.0

# File Processing
aiofiles==23.2.1

# Logging
loguru==0.7.2
